small refactors & resolve comments
